chore: move toggle groups to pet register folder instead of Custom folder
